logic Merge time and date module into one datetime for simplicity
