+++ logic/src/lib.rs
@@ -2,6 +2,5 @@
 
 #![deny(missing_docs)] // Logic is a main shared library - require docs for all public interfaces
 
-pub mod date;
+pub mod datetime;
 pub mod messages;
-pub mod time;
